♻️ refactor: setup 파일 한 디렉토리에서 관리하도록 경로 수정
